+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Go string representation of `Identifier` is now a string.
+
 ## [0.4.0] - 2024-09-06
 
 ### Changed